feat(range-slider): replace legacy slider with new one
