+++ dev/index.html
@@ -30,7 +30,6 @@
         <div class="facet" id="free-shipping"></div>
         <div class="facet" id="google-amazon"></div>
         <div class="facet" id="price"></div>
-        <div class="facet" id="price2"></div>
         <div class="facet" id="categories"></div>
         <div class="facet" id="price-ranges"></div>
         <div class="facet" id="price-numeric-list"></div>
