+++ src/components/Slider/Pit.js
