--- src/components/Slider2/__tests__/Slider-test.js
@@ -1,22 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-
-import Slider from '../Slider2';
-
-describe('Slider', () => {
-  it('should render correctly', () => {
-    const tree = renderer.create(
-      <Slider
-        refine={ () => undefined }
-        min={ 0 }
-        max={ 500 }
-        values={ [0, 0] }
-        pips={ true }
-        step={ 2 }
-        tooltips={ true }
-        shouldAutoHideContainer={ false }
-      />
-    ).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-});
